backend/pyproject.toml
[project]
name = "docdelta-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Satwik Kapavarapu",email = "satwiksakethkapavarapu@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pdfminer-six (==20221105)",
    "python-docx (>=1.1.2,<2.0.0)",
    "diff-match-patch (>=20241021,<20241022)",
    "protobuf (>=6.31.1,<7.0.0)",
    "grpcio (>=1.72.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
ruff = "^0.11.12"
black = "^25.1.0"
pre-commit = "^4.2.0"


[tool.black]
line-length = 88

[tool.ruff]
line-length = 88
